requirements.txt
streamlit
langchain
langgraph
imap-tools
python-dotenv
ollama